+++ src/source/protocol_crud.ts
@@ -1,0 +1,13 @@
+import {AddProtocolAction, config} from '../util/config';
+
+export function getProtocol(url: string) {
+    return config.REGISTERED_PROTOCOLS[url.substring(0, url.indexOf('://'))];
+}
+
+export function addProtocol(customProtocol: string, loadFn: AddProtocolAction) {
+    config.REGISTERED_PROTOCOLS[customProtocol] = loadFn;
+}
+
+export function removeProtocol(customProtocol: string) {
+    delete config.REGISTERED_PROTOCOLS[customProtocol];
+}