+++ src/source/worker.ts
@@ -5,6 +5,7 @@ import {RasterDEMTileWorkerSource} from './raster_dem_tile_worker_source';
 import {rtlWorkerPlugin, RTLTextPlugin} from './rtl_text_plugin_worker';
 import {GeoJSONWorkerSource, LoadGeoJSONParameters} from './geojson_worker_source';
 import {isWorker} from '../util/util';
+import {addProtocol, removeProtocol} from './protocol_crud';
 
 import type {
     WorkerSource,
@@ -71,6 +72,9 @@ export default class Worker {
             }
             this.externalWorkerSourceTypes[name] = WorkerSource;
         };
+
+        this.self.addProtocol = addProtocol;
+        this.self.removeProtocol = removeProtocol;
 
         // This is invoked by the RTL text plugin when the download via the `importScripts` call has finished, and the code has been parsed.
         this.self.registerRTLTextPlugin = (rtlTextPlugin: RTLTextPlugin) => {
@@ -143,7 +147,7 @@ export default class Worker {
             return this._syncRTLPluginState(mapId, params);
         });
 
-        this.actor.registerMessageHandler('loadWorkerSource', async (_mapId: string, params: string) => {
+        this.actor.registerMessageHandler('importScript', async (_mapId: string, params: string) => {
             this.self.importScripts(params);
         });