+++ src/util/web_worker.ts
@@ -1,4 +1,4 @@
-import {config} from './config';
+import {AddProtocolAction, config} from './config';
 import type {default as MaplibreWorker} from '../source/worker';
 import type {WorkerSourceConstructor} from '../source/worker_source';
 
@@ -6,6 +6,8 @@ export interface WorkerGlobalScopeInterface {
     importScripts(...urls: Array<string>): void;
     registerWorkerSource: (sourceName: string, sourceConstrucor: WorkerSourceConstructor) => void;
     registerRTLTextPlugin: (_: any) => void;
+    addProtocol: (customProtocol: string, loadFn: AddProtocolAction) => void;
+    removeProtocol: (customProtocol: string) => void;
     worker: MaplibreWorker;
 }